server/register-rest/requirements.txt
Django==1.5.1
gunicorn==18.0
django-tastypie==0.10.0
python-mimeparse==0.1.4
MySQL-python==1.2.5
pygeoip==0.2.5
argparse==1.2.1
python-dateutil==2.2
pytz==2014.2
six==1.6.1
wsgiref==0.1.2
